api/requirements.txt
fastapi[standard]
openai-whisper
requests
sqlalchemy
python-dotenv
moviepy
imageio[ffmpeg]
scikit-learn